BFF for streams connected
